backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
python-multipart==0.0.6
pydantic==2.5.2
Pillow==10.0.1
pytesseract==0.3.10
jinja2==3.1.2
requests

